+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added `CompositePropagator` that accepts a list of `ITextFormat` following
+  [specification](https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/context/api-propagators.md#create-a-composite-propagator)
+
 ## 0.3.0-beta
 
 Released 2020-07-23